requirements.txt
# Core Audio Processing
librosa==0.11.0
soundfile==0.13.1
audioread==3.1.0

# Spotify API
spotipy>=2.23.0

# OpenAI API
openai>=1.0.0

# HTTP & API
requests>=2.31.0

# Environment Variables
python-dotenv>=1.0.0

# Audio File Metadata
mutagen>=1.47.0

# Scientific Computing
numpy>=1.24.0
scipy>=1.11.0
numba==0.63.1
llvmlite==0.46.0

# Progress Bars
tqdm>=4.66.0

# Data Validation (optional)
pydantic>=2.0.0

# Development Tools
pre-commit>=3.5.0

# Testing (optional - uncomment if needed)
# pytest>=7.4.0
# pytest-cov>=4.1.0

psutil
